Fix 11 failing E2E tests for provider-aware model selection
The ModelSelector component now uses providerId::modelId format for
option values instead of bare model IDs. Updated all model selector
tests to use prefixed format. Also added required summaryModel and
summaryProviderId fields to summary settings API calls.
