+++ packages/server/src/db/ProjectDefaultsRepository.js
@@ -74,48 +74,53 @@ export class ProjectDefaultsRepository extends BaseRepository {
    * @param {Object} data - Defaults data (all fields optional)
    * @returns {Object} Updated defaults object
    */
+  #insertDefaults(projectId, data) {
+    const id = databaseManager.generateId();
+    const now = Date.now();
+
+    this.db
+      .prepare(
+        `INSERT INTO project_session_defaults
+         (id, project_id, mode, thinking_enabled, start_immediately, git_mode, git_branch, model, provider_id, effort_level, created_at, updated_at)
+         VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
+      )
+      .run(
+        id,
+        projectId,
+        data.mode || null,
+        data.thinkingEnabled !== undefined ? (data.thinkingEnabled ? 1 : 0) : null,
+        data.startImmediately !== undefined ? (data.startImmediately ? 1 : 0) : null,
+        data.gitMode || null,
+        data.gitBranch || null,
+        data.model || null,
+        data.providerId || null,
+        data.effortLevel || null,
+        now,
+        now
+      );
+  }
+
+  #updateDefaults(projectId, data) {
+    const { updates, values } = buildUpdateFields(data);
+
+    if (updates.length > 0) {
+      updates.push('updated_at = ?');
+      values.push(Date.now());
+      values.push(projectId);
+
+      this.db
+        .prepare(`UPDATE project_session_defaults SET ${updates.join(', ')} WHERE project_id = ?`)
+        .run(...values);
+    }
+  }
+
   upsert(projectId, data) {
-    // Get existing defaults if any
     const existing = this.getByProjectId(projectId);
 
     if (!existing) {
-      // Create new defaults record
-      const id = databaseManager.generateId();
-      const now = Date.now();
-
-      this.db
-        .prepare(
-          `INSERT INTO project_session_defaults
-           (id, project_id, mode, thinking_enabled, start_immediately, git_mode, git_branch, model, provider_id, effort_level, created_at, updated_at)
-           VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-        )
-        .run(
-          id,
-          projectId,
-          data.mode || null,
-          data.thinkingEnabled !== undefined ? (data.thinkingEnabled ? 1 : 0) : null,
-          data.startImmediately !== undefined ? (data.startImmediately ? 1 : 0) : null,
-          data.gitMode || null,
-          data.gitBranch || null,
-          data.model || null,
-          data.providerId || null,
-          data.effortLevel || null,
-          now,
-          now
-        );
+      this.#insertDefaults(projectId, data);
     } else {
-      // Update existing defaults
-      const { updates, values } = buildUpdateFields(data);
-
-      if (updates.length > 0) {
-        updates.push('updated_at = ?');
-        values.push(Date.now());
-        values.push(projectId);
-
-        this.db
-          .prepare(`UPDATE project_session_defaults SET ${updates.join(', ')} WHERE project_id = ?`)
-          .run(...values);
-      }
+      this.#updateDefaults(projectId, data);
     }
 
     return this.getByProjectId(projectId);
